requirements.txt
exifread
folium
pillow
piexif
